refactor: second-pass shared code cleanup
Fix GameItem.tsx getWinner bug (missed "post" state, same as GameChip
fix from first pass). Extract shared utilities and consolidate:

- gameStatusLabel() added to gameHelpers.ts, replaces chipStatus(),
  gameStatus(), and statusLabel() in 3 files
- useScoreFlash hook extracted from GameItem, replaces inline flash
  logic in GameChip
- formatUptime() and formatPollInterval() added to utils/format.ts,
  replaces duplicates in sysmon/FeedTab and SysmonSummary
- usageColor() moved to sysmon/utils.ts, replaces duplicate in
  SysmonSummary
- formatNextGame() deleted from SportsConfigPanel, reuses
  formatCountdown() from gameHelpers
- Custom CloseIcon SVGs replaced with lucide X in WorldClock and
  WeatherCard
