+++ desktop/src/utils/format.ts
@@ -91,6 +91,32 @@ export function truncate(text: string, maxLen: number): string {
 }
 
 /**
+ * Format a duration in seconds as human-readable uptime.
+ * Returns "2d 5h", "3h 12m", or "45m".
+ */
+export function formatUptime(secs: number): string {
+  const days = Math.floor(secs / 86400);
+  const hours = Math.floor((secs % 86400) / 3600);
+  const mins = Math.floor((secs % 3600) / 60);
+  if (days > 0) return `${days}d ${hours}h`;
+  if (hours > 0) return `${hours}h ${mins}m`;
+  return `${mins}m`;
+}
+
+/**
+ * Format a poll interval in seconds as human-readable duration.
+ * Returns "2m", "2m 30s", or "30s".
+ */
+export function formatPollInterval(secs: number): string {
+  if (secs >= 60) {
+    const m = Math.floor(secs / 60);
+    const s = secs % 60;
+    return s ? `${m}m ${s}s` : `${m}m`;
+  }
+  return `${secs}s`;
+}
+
+/**
  * Format a byte count as a human-readable string.
  * Handles B, KB, MB, GB, and TB.
  */
