+++ desktop/src/widgets/clock/WorldClock.tsx
@@ -4,6 +4,7 @@
  * Timezone selections and format (12h/24h) are persisted to Tauri store.
  */
 import { useState, useEffect, useCallback, useRef, useMemo } from "react";
+import { X } from "lucide-react";
 import Tooltip from "../../components/Tooltip";
 import { getStore, setStore } from "../../lib/store";
 import type { TimeFormat, TimezoneEntry } from "./types";
@@ -136,26 +137,6 @@ function getLocalLabel(): string {
   }
 }
 
-// ── Close icon ──────────────────────────────────────────────────
-
-function CloseIcon({ size = 12 }: { size?: number }) {
-  return (
-    <svg
-      width={size}
-      height={size}
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2.5"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      aria-hidden="true"
-    >
-      <path d="M18 6L6 18M6 6l12 12" />
-    </svg>
-  );
-}
-
 // ── Clock Card ──────────────────────────────────────────────────
 
 function ClockCard({
@@ -216,7 +197,7 @@ function ClockCard({
                 onClick={onRemove}
                 className="text-fg-3 hover:text-error opacity-0 group-hover:opacity-100 transition-all"
               >
-                <CloseIcon size={11} />
+                <X size={11} />
               </button>
             </Tooltip>
           )}
@@ -245,7 +226,7 @@ function ClockCard({
             onClick={onRemove}
             className="absolute top-2.5 right-2.5 w-5 h-5 flex items-center justify-center rounded text-fg-3 hover:text-error hover:bg-error/10 opacity-0 group-hover:opacity-100 transition-all"
           >
-            <CloseIcon size={11} />
+            <X size={11} />
           </button>
         </Tooltip>
       )}
